⬆️ (package.json): upgrade NX workspace and related packages to 19.6.3 ✨ (package.json): add "migrate" script for easier NX migrations
Upgrading NX workspace and its related packages ensures that the project stays up-to-date with the latest features, bug fixes, and performance improvements provided by the NX tooling. The addition of the "migrate" script simplifies the process of updating NX itself and its plugins to the latest versions, enhancing the development workflow and maintaining compatibility with new changes.