✨ (episodes-container.component.ts): add SkeletonComponent to improve UX during data loading ♻️ (episodes-item.component.ts): remove unused OnInit lifecycle and outputs for cleaner code ✅ (skeleton.component.spec.ts): add tests for SkeletonComponent to ensure reliability ✨ (skeleton.component.ts): introduce SkeletonComponent for a better loading state representation 🔧 (loading.interceptor.ts, loading.service.ts): refine loading service and interceptor for enhanced debuggability and code clarity
Introducing the SkeletonComponent enhances the user experience by providing a more visually appealing loading state. Cleaning up the EpisodesItemComponent by removing unused code makes the component simpler and more focused on its purpose. Adding tests for the new SkeletonComponent ensures that it works as expected and maintains the quality of the application. Refinements in the loading service and interceptor improve the debuggability of loading states and streamline the codebase for better maintainability.