📝 (articles): add new article on testing in Angular and related assets
Introduce a new article titled "Testing in Angular: From Unit to
Integration Testing" to explore modern testing practices in Angular
applications. The article discusses the shift from traditional unit
testing to integration testing, based on insights from Rainer Hanekamp.
It aims to provide developers with a comprehensive understanding of
testing strategies and encourages the adoption of more holistic testing
approaches. Additionally, a cover image is added to enhance the visual
appeal of the article.