📝 (articles): publish NgRx Signal Store article and add update article
The NgRx Signal Store article is now published to share insights on state management in Angular. An update article is added to highlight recent enhancements in NgRx Signal Store, focusing on improved modularity and alignment with best practices. This aims to provide readers with the latest information and techniques for efficient state management.

⬆️ (package.json): upgrade @ngrx/signals to version 19.0.1

The @ngrx/signals package is upgraded to version 19.0.1 to incorporate the latest features and improvements. This ensures compatibility with the updated NgRx Signal Store implementation and leverages new functionalities for better state management.

♻️ (podcast.store.ts): refactor to use withProps for dependency injection

The PodcastStore is refactored to use `withProps` for injecting dependencies like PodcastService. This change centralizes dependency management, reduces redundancy, and aligns with NgRx documentation, enhancing code readability and maintainability.